major refactoring. examples work. still need to update documentation
